requirements.txt
mlflow
pandas
scikit-learn
joblib
azure-identity
azure-ai-ml